Remove old jose code (#186)
